WEBAPP-223: Remove old withFetcher, endpoint stuff
