+++ src/modules/app/containers/App.js
@@ -7,7 +7,7 @@ import I18nProvider from './I18nProvider'
 import RouteConfig from '../RouteConfig'
 import RouterFragment from './RouterFragment'
 import createRouteConfig from '../createRouteConfig'
-import withFetcher from '../../endpoint/withFetcher'
+import withFetcher from '../../endpoint/hocs/withFetcher'
 import locationsMapper from '../../endpoint/mappers/locations'
 import { locationUrlMapper } from '../../endpoint/urls'
