+++ src/modules/app/containers/LanguageFragment.js
@@ -13,7 +13,7 @@ import RouteConfig from '../RouteConfig'
 import { connect } from 'react-redux'
 import LocationModel from '../../endpoint/models/LocationModel'
 import LanguageModel from '../../endpoint/models/LanguageModel'
-import withFetcher from '../../endpoint/withFetcher'
+import withFetcher from '../../endpoint/hocs/withFetcher'
 import compose from 'lodash/fp/compose'
 import EventModel from '../../endpoint/models/EventModel'
 import ExtraModel from '../../endpoint/models/ExtraModel'
