+++ src/modules/app/containers/LocationFragment.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types'
 import RouteConfig from '../RouteConfig'
 import { connect } from 'react-redux'
 import LanguageModel from '../../endpoint/models/LanguageModel'
-import withFetcher from '../../endpoint/withFetcher'
+import withFetcher from '../../endpoint/hocs/withFetcher'
 import LocationModel from '../../endpoint/models/LocationModel'
 import LanguageFragment from './LanguageFragment'
 import LanguageFailure from '../../../routes/categories/containers/LanguageFailure'
