+++ src/modules/app/createReduxStore.js
@@ -4,7 +4,6 @@ import { initializeCurrentLocation, routerForBrowser } from 'redux-little-router
 import { createLogger } from 'redux-logger'
 
 import RouteConfig from 'modules/app/RouteConfig'
-import endpointReducers from 'modules/endpoint/reducers'
 import setLanguageChangeUrlsReducer from '../language/reducers/setLanguageChangeUrls'
 import { createResponsiveStateReducer, responsiveStoreEnhancer } from 'redux-responsive'
 
@@ -33,7 +32,6 @@ const createReduxStore = (createHistory, initialState = {}, routes = new RouteCo
   }
 
   const reducer = combineReducers({
-    ...endpointReducers,
     viewport: createResponsiveStateReducer({small: 750}, {infinity: 'large'}),
     router: routerReducer,
     languageChangeUrls: setLanguageChangeUrlsReducer
