--- src/modules/endpoint/__tests__/EndpointProvider.spec.js
@@ -1,23 +1,0 @@
-import * as React from 'react'
-import { shallow } from 'enzyme'
-import EndpointProvider from '../EndpointProvider'
-import EndpointBuilder from '../EndpointBuilder'
-
-describe('EndpointProvider', () => {
-  const endpoint = new EndpointBuilder('endpoint')
-    .withStateToUrlMapper(() => 'https://weird-endpoint/api.json')
-    .withMapper(json => json)
-    .withResponseOverride({})
-    .build()
-
-  it('should have correct child context', () => {
-    expect(EndpointProvider.childContextTypes).toHaveProperty('getEndpoint')
-
-    const provider = shallow(<EndpointProvider endpoints={[endpoint]}>
-      <div />
-    </EndpointProvider>).instance()
-
-    const getEndpoint = provider.getChildContext()['getEndpoint']
-    expect(getEndpoint(endpoint.stateName)).toBe(endpoint)
-  })
-})
