--- src/modules/endpoint/reducers/index.js
@@ -1,26 +1,0 @@
-import languagesEndpoint from '../endpoints/languages'
-import locationEndpoint from '../endpoints/locations'
-import categoriesEndpoint from '../endpoints/categories'
-import eventsEndpoint from '../endpoints/events'
-import disclaimerEndpoint from '../endpoints/disclaimer'
-import extrasEndpoint from '../endpoints/extras'
-import sprungbrettEndpoint from '../endpoints/sprungbrett'
-
-/**
- * Contains all reducers from all endpoints which are defined in {@link './endpoints/'}
- */
-const endpoints = [
-  languagesEndpoint,
-  locationEndpoint,
-  categoriesEndpoint,
-  disclaimerEndpoint,
-  eventsEndpoint,
-  extrasEndpoint,
-  sprungbrettEndpoint
-]
-const reducers = endpoints.reduce((result, endpoint) => {
-  result[endpoint.stateName] = endpoint.createReducer()
-  return result
-}, {})
-
-export default reducers
